:recycle: refactor depdency injections
