+++ src/Mindee/MindeeClient.cs
@@ -4,8 +4,8 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
-using Mindee.DependencyInjection;
 using Mindee.Exceptions;
+using Mindee.Extensions.DependencyInjection;
 using Mindee.Http;
 using Mindee.Input;
 using Mindee.Parsing.Common;
