+++ tests/Mindee.UnitTests/UnitTestBase.cs
@@ -1,7 +1,7 @@
 using System.Net;
 using System.Net.Http; // Necessary for .NET 4.7.2
 using Microsoft.Extensions.DependencyInjection;
-using Mindee.DependencyInjection;
+using Mindee.Extensions.DependencyInjection;
 using Mindee.Http;
 using Mindee.Input;
 using Moq;
